Rename provider_base.py to base.py, for consistency with executors and channels
This will break external implementations of providers which rely directly on
importing provider_base, but I am not aware of any at the time of writing.
