+++ docs/reference.rst
@@ -119,7 +119,7 @@ Providers
     parsl.providers.TorqueProvider
     parsl.providers.KubernetesProvider
     parsl.providers.PBSProProvider
-    parsl.providers.provider_base.ExecutionProvider
+    parsl.providers.base.ExecutionProvider
     parsl.providers.cluster_provider.ClusterProvider
