+++ mypy.ini
@@ -21,7 +21,7 @@ ignore_missing_imports = True
 # parsl to be checked that way, so these options could be default instead of
 # listed per package; but unless/until that happens, the more-strictly-checked
 # code should be listed here:
-[mypy-parsl.providers.provider_base.*]
+[mypy-parsl.providers.base.*]
 no_implicit_optional = True
 disallow_untyped_decorators = True
 warn_unused_ignores = True
