+++ parsl/dataflow/strategy.py
@@ -6,7 +6,7 @@ from typing import List
 
 from parsl.dataflow.executor_status import ExecutorStatus
 from parsl.executors import HighThroughputExecutor
-from parsl.providers.provider_base import JobState
+from parsl.providers.base import JobState
 from parsl.process_loggers import wrap_with_logs
