+++ parsl/executors/extreme_scale/executor.py
@@ -60,7 +60,7 @@ class ExtremeScaleExecutor(HighThroughputExecutor, RepresentationMixin):
     Parameters
     ----------
 
-    provider : :class:`~parsl.providers.provider_base.ExecutionProvider`
+    provider : :class:`~parsl.providers.base.ExecutionProvider`
        Provider to access computation resources. Can be any providers in ``parsl.providers``:
         :class:`~parsl.providers.cobalt.cobalt.Cobalt`,
         :class:`~parsl.providers.condor.condor.Condor`,
