+++ parsl/providers/azure/azure.py
@@ -6,7 +6,7 @@ from string import Template
 
 from parsl.dataflow.error import ConfigurationError
 from parsl.providers.azure.template import template_string
-from parsl.providers.provider_base import ExecutionProvider, JobState, JobStatus
+from parsl.providers.base import ExecutionProvider, JobState, JobStatus
 from parsl.errors import OptionalModuleMissing
 from parsl.utils import RepresentationMixin
 from parsl.launchers import SingleNodeLauncher
