+++ parsl/providers/googlecloud/googlecloud.py
@@ -2,7 +2,7 @@ import atexit
 import logging
 import os
 from parsl.launchers import SingleNodeLauncher
-from parsl.providers.provider_base import JobState, JobStatus
+from parsl.providers.base import JobState, JobStatus
 
 logger = logging.getLogger(__name__)
