+++ parsl/providers/kubernetes/kube.py
@@ -5,7 +5,7 @@ from parsl.providers.kubernetes.template import template_string
 logger = logging.getLogger(__name__)
 
 from parsl.errors import OptionalModuleMissing
-from parsl.providers.provider_base import ExecutionProvider, JobState, JobStatus
+from parsl.providers.base import ExecutionProvider, JobState, JobStatus
 from parsl.utils import RepresentationMixin
 
 import typeguard
