+++ parsl/providers/pbspro/pbspro.py
@@ -7,7 +7,7 @@ from parsl.channels import LocalChannel
 from parsl.launchers import SingleNodeLauncher
 from parsl.providers.pbspro.template import template_string
 from parsl.providers import TorqueProvider
-from parsl.providers.provider_base import JobState, JobStatus
+from parsl.providers.base import JobState, JobStatus
 
 from parsl.providers.torque.torque import translate_table
